pyproject.toml
[project]
name = "analiza-metody-techniki"
version = "1.0.0"
description = "Likert Survey Analysis Pipeline for budget priorities and security research"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "seaborn>=0.13.0",
    "statsmodels>=0.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
]
llm = [
    "openai>=1.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
